Creation of common files for itsme and itsModel
It keep similar functions always updated in the same way
